Move `CodableFeedStore` into a separate file
