+++ Tests/FeedStoreChallengeTests.swift
@@ -5,81 +5,6 @@
 import XCTest
 import FeedStoreChallenge
 
-class LocalFeedStore: FeedStore {
-	private struct Cache: Codable {
-		let feed: [CodableFeedImage]
-		let timestamp: Date
-		
-		var toLocal: [LocalFeedImage] {
-			feed.map { $0.local }
-		}
-	}
-	
-	private struct CodableFeedImage: Codable {
-		public let id: UUID
-		public let description: String?
-		public let location: String?
-		public let url: URL
-		
-		init(_ image: LocalFeedImage) {
-			self.id = image.id
-			self.description = image.description
-			self.location = image.location
-			self.url = image.url
-		}
-		
-		var local: LocalFeedImage {
-			LocalFeedImage(id: id, description: description, location: location, url: url)
-		}
-	}
-	
-	let storeURL: URL
-	
-	init(storeURL: URL) {
-		self.storeURL = storeURL
-	}
-	
-	func deleteCachedFeed(completion: @escaping DeletionCompletion) {
-		guard FileManager.default.fileExists(atPath: storeURL.path) else {
-			return completion(nil)
-		}
-		do {
-			try FileManager.default.removeItem(at: storeURL)
-			completion(nil)
-		} catch {
-			completion(error)
-		}
-	}
-	
-	func insert(_ feed: [LocalFeedImage], timestamp: Date, completion: @escaping InsertionCompletion) {
-		do {
-			let encoder = JSONEncoder()
-			let cache = Cache(feed: feed.map { CodableFeedImage($0) }, timestamp: timestamp)
-			let data = try encoder.encode(cache)
-			try data.write(to: storeURL)
-		} catch {
-			return completion(error)
-		}
-		completion(nil)
-	}
-	
-	func retrieve(completion: @escaping RetrievalCompletion) {
-		guard let data = try? Data(contentsOf: storeURL) else {
-			return completion(.empty)
-		}
-		
-		do {
-			let decoder = JSONDecoder()
-			let cache = try decoder.decode(Cache.self, from: data)
-			completion(.found(feed: cache.toLocal, timestamp: cache.timestamp))
-		} catch {
-			completion(.failure(error))
-		}
-	}
-	
-	
-}
-
 class FeedStoreChallengeTests: XCTestCase, FeedStoreSpecs {
 	
 	//  ***********************
@@ -181,7 +106,7 @@ class FeedStoreChallengeTests: XCTestCase, FeedStoreSpecs {
 	// - MARK: Helpers
 	
 	private func makeSUT(url: URL? = nil) -> FeedStore {
-		return LocalFeedStore(storeURL: url ?? testSpecificStoreURL())
+		return CodableFeedStore(storeURL: url ?? testSpecificStoreURL())
 	}
 	
 	private func testSpecificStoreURL() -> URL {
